moved lib into classes/local for autoloading purposes
