+++ classes/local/group.php
@@ -14,7 +14,7 @@
 // You should have received a copy of the GNU General Public License
 // along with Moodle.  If not, see <http://www.gnu.org/licenses/>.
 
-namespace ratingallocate;
+namespace mod_ratingallocate\local;
 
 class group {
