+++ classes/local/lp/engine.php
@@ -14,7 +14,26 @@
 // You should have received a copy of the GNU General Public License
 // along with Moodle.  If not, see <http://www.gnu.org/licenses/>.
 
-require_once(dirname(__FILE__).'/executor.php');
-require_once(dirname(__FILE__).'/executors/local.php');
-require_once(dirname(__FILE__).'/executors/ssh.php');
-require_once(dirname(__FILE__).'/executors/webservice.php');
+namespace mod_ratingallocate\local\lp;
+
+abstract class engine {
+
+    /**
+     * Reads the content of the stream and returns the variables and their optimized values
+     *
+     * @param $stream Output stream of the program that was executed
+     *
+     * @return Array of variables and their values
+     */
+    abstract public function read($stream);
+
+    /**
+     * Returns the command that gets executed
+     *
+     * @param $input_file Name of the input file
+     *
+     * @returns Command as a string
+     */
+    abstract public function get_command($input_file);
+
+};
