+++ classes/local/lp/linear_program.php
@@ -14,7 +14,7 @@
 // You should have received a copy of the GNU General Public License
 // along with Moodle.  If not, see <http://www.gnu.org/licenses/>.
 
-namespace ratingallocate\lp;
+namespace mod_ratingallocate\local\lp;
 
 class linear_program {
     const NONE = -1;
@@ -25,7 +25,7 @@ class linear_program {
     const INTEGER = 1;
     const REAL = 2;
     const COMPLEX = 3;
-    
+
     private $objective_method = self::NONE;
     private $objective_function = '';
 
@@ -44,7 +44,7 @@ class linear_program {
     public function set_objective_function($objective_function) {
         $this->objective_function = $objective_function;
     }
-    
+
     public function get_objective_function() {
         return $this->objective_function;
     }
@@ -55,25 +55,25 @@ class linear_program {
     }
 
     public function set_constraints($constraints) {
-        $this->constraints = $constraints; 
+        $this->constraints = $constraints;
     }
 
     public function get_constraints() {
         return $this->constraints;
     }
-    
+
     public function add_constraint($constraint) {
         $this->constraints[] = $constraint;
     }
 
     public function set_bounds($bounds) {
-        $this->bounds = $bounds; 
+        $this->bounds = $bounds;
     }
 
     public function get_bounds() {
         return $this->bounds;
     }
-    
+
     public function add_bound($bound) {
         $this->bounds[] = $bound;
     }
@@ -81,11 +81,11 @@ class linear_program {
     public function set_variables($variables) {
         $this->variables = $variables;
     }
-   
+
     public function get_variables() {
         return $this->variables;
     }
-    
+
     public function add_variable($variable, $type = self::REAL) {
         $this->variables[$variable] = ['type' => $type, 'name' => $variable];
     }
@@ -98,32 +98,32 @@ class linear_program {
 
         throw new \exception('Linear program objectives method is invalid!');
     }
-    
+
     public function write_objective() {
         if(empty($this->objective_function))
             throw new \exception('Linear program objectives function is invalid!');
 
         return $this->write_objective_method($this->objective_method)."\n  ".str_replace('*', ' ', $this->objective_function)."\n";
     }
-    
+
     public function write_constraints() {
         if(empty($this->constraints))
             return '';
-        
+
         return "Subject To\n".implode(array_map(function($constraint) {
             return "  $constraint\n";
-        }, $this->constraints));        
+        }, $this->constraints));
     }
 
     public function write_bounds() {
         if(empty($this->bounds))
             return '';
-        
+
         return "Bounds\n".implode(array_map(function($bound) {
             return "  $bound\n";
         }, $this->bounds));
     }
-    
+
     public function write_variables() {
         if(empty($this->variables))
             return '';
@@ -132,7 +132,7 @@ class linear_program {
             return $variable['name'].' ';
         }, $this->variables))."\n";
     }
-    
+
     public function write() {
         return $this->write_objective()
             .$this->write_constraints()
