+++ classes/local/lp/weighters/identity_weighter.php
@@ -14,13 +14,13 @@
 // You should have received a copy of the GNU General Public License
 // along with Moodle.  If not, see <http://www.gnu.org/licenses/>.
 
-namespace ratingallocate\lp\weighters;
+namespace mod_ratingallocate\local\lp\weighters;
 
 /**
  * Class which represents an identity weighter
  */
 class identity_weighter extends polynomial_weighter {
-    
+
     /**
      * Creates an identity weighter
      */
@@ -29,4 +29,3 @@ class identity_weighter extends polynomial_weighter {
     }
 
 };
-
