--- lib/lp/engine.php
@@ -1,39 +1,0 @@
-<?php
-// This file is part of Moodle - http://moodle.org/
-//
-// Moodle is free software: you can redistribute it and/or modify
-// it under the terms of the GNU General Public License as published by
-// the Free Software Foundation, either version 3 of the License, or
-// (at your option) any later version.
-//
-// Moodle is distributed in the hope that it will be useful,
-// but WITHOUT ANY WARRANTY; without even the implied warranty of
-// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-// GNU General Public License for more details.
-//
-// You should have received a copy of the GNU General Public License
-// along with Moodle.  If not, see <http://www.gnu.org/licenses/>.
-
-namespace ratingallocate\lp;
-
-abstract class engine {
-    
-    /**
-     * Reads the content of the stream and returns the variables and their optimized values
-     *
-     * @param $stream Output stream of the program that was executed
-     *
-     * @return Array of variables and their values
-     */
-    abstract public function read($stream);
-
-    /**
-     * Returns the command that gets executed
-     *
-     * @param $input_file Name of the input file
-     *
-     * @returns Command as a string 
-     */
-    abstract public function get_command($input_file);
-
-};
