+++ solver/lp-solver.php
@@ -28,15 +28,9 @@ defined('MOODLE_INTERNAL') || die();
 require_once(dirname(__FILE__) . '/../locallib.php');
 require_once(dirname(__FILE__) . '/solver-template.php');
 
-require_once(dirname(__FILE__).'/../lib/user.php');
-require_once(dirname(__FILE__).'/../lib/group.php');
-require_once(dirname(__FILE__).'/../lib/ssh.php');
-require_once(dirname(__FILE__).'/../lib/lp.php');
-require_once(dirname(__FILE__).'/../lib/utility.php');
-
-function dbg_msg($obj) {
-    echo "<pre>", print_r($obj), "</pre>";
-}
+//function dbg_msg($obj) {
+//    echo "<pre>", print_r($obj), "</pre>";
+//}
 
 class solver_lp extends distributor {
 
@@ -47,44 +41,49 @@ class solver_lp extends distributor {
     public function get_ssh_connection() {
         global $CFG;
 
-        return new \ratingallocate\ssh\connection($CFG->ratingallocate_ssh_hostname,
-                                                  new \ratingallocate\ssh\password_authentication($CFG->ratingallocate_ssh_username,
-                                                                                                  $CFG->ratingallocate_ssh_password));
+        return new \mod_ratingallocate\local\ssh\connection($CFG->ratingallocate_ssh_hostname,
+                                                            $this->get_ssh_authentication());
     }
-    
+
+    public function get_ssh_authentication() {
+        global $CFG;
+
+        return new \mod_ratingallocate\local\ssh\password_authentication($CFG->ratingallocate_ssh_username,
+                                                                         $CFG->ratingallocate_ssh_password);
+    }
+
     public function get_executor($engine) {
         global $CFG;
 
         switch($CFG->ratingallocate_executor)
         {
         case 'webservice':
-            return new \ratingallocate\lp\executors\webservice\connector($engine, $CFG->ratingallocate_uri, $CFG->ratingallocate_secret);
-            
+            return new \mod_ratingallocate\local\lp\executors\webservice\connector($engine, $CFG->ratingallocate_uri, $CFG->ratingallocate_secret);
+
         case 'ssh':
-            return new \ratingallocate\lp\executors\ssh($engine, $this->get_ssh_connection(), $CFG->ratingallocate_remote_path);
+            return new \mod_ratingallocate\local\lp\executors\ssh($engine, $this->get_ssh_connection(), $CFG->ratingallocate_remote_path);
         }
 
-        return new \ratingallocate\lp\executors\local($engine, $CFG->ratingallocate_local_path);
+        return new \mod_ratingallocate\lp\local\executors\local($engine, $CFG->ratingallocate_local_path);
     }
 
     public function get_engine() {
         global $CFG;
 
-        $engine_path = "\\ratingallocate\\lp\\engines\\{$CFG->ratingallocate_engine}";
+        $engine_path = "\\mod_ratingallocate\\local\\lp\\engines\\{$CFG->ratingallocate_engine}";
+
         return new $engine_path();
     }
 
     public function compute_distribution($choicerecords, $ratings, $usercount) {
-        $configuration = \ratingallocate\utility::get_configuration($CFG);
-
-        list($users, $groups) = \ratingallocate\utility::transform_to_users_and_groups($choicerecords, $ratings);
+        list($users, $groups) = \mod_ratingallocate\local\utility::transform_to_users_and_groups($choicerecords, $ratings);
 
         $engine = $this->get_engine();
         $executor = $this->get_executor($engine);
-        
-        $executor->solve_objects($users, $groups, new \ratingallocate\lp\weighters\identity_weighter());
 
-        return \ratingallocate\utility::transform_from_users_and_groups($users, $groups);
+        $executor->solve_objects($users, $groups, new \mod_ratingallocate\local\lp\weighters\identity_weighter());
+
+        return \mod_ratingallocate\local\utility::transform_from_users_and_groups($users, $groups);
     }
 
 }
